requirements.txt
flask==3.0.0
dashscope==1.14.1
PyMuPDF==1.23.8
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
werkzeug==3.0.1
